Add case for special moment settings
My daily note format "[6. ledger Productivity]/[Journal]/YYYY/MM - MMMM/YYYY-MM-DD" wihle valid in Moment and in Obsidian settings created issue with repeated tasks because the current way was just looking for the final end (YYYY-MM-DD) so it was never valid.  I switched it to look for both case and it doesn't seems to have created problems so far in my use so I'm sending it if you want to integrate for possibly others who have the same sort of settings.